Consume .proto file from separate repo and reflect enableContentNegotiation change (#59)
* Added subtree from https://github.com/azure/azure-functions-language-worker-protobuf. Branch: dev. Release: v1.0.0

* Consume new proto file from subtree

* remove package-lock changes

* remove references to raw
