+++ package.json
@@ -26,12 +26,12 @@
     "webpack": "^3.5.6"
   },
   "scripts": {
-    "clean": "rimraf dist && rimraf protos/rpc*",
-    "build": "rimraf dist && npm run gen && shx mkdir -p dist/protos && shx cp protos/rpc.* dist/protos/. && tsc",
+    "clean": "rimraf dist && rimraf azure-functions-language-worker-protobuf/src/rpc*",
+    "build": "rimraf dist && npm run gen && shx mkdir -p dist/protos && shx cp azure-functions-language-worker-protobuf/src/rpc.* dist/protos/. && tsc",
     "build-nomaps": "npm run build -- --sourceMap false",
     "gen": "npm run genJs && npm run genTs",
-    "genJs": "pbjs -t json-module -w commonjs -o protos/rpc.js protos/FunctionRpc.proto",
-    "genTs": "pbjs -t static-module protos/FunctionRpc.proto -o protos/rpc_static.js && pbts -o protos/rpc.d.ts protos/rpc_static.js",
+    "genJs": "pbjs -t json-module -w commonjs -o azure-functions-language-worker-protobuf/src/rpc.js azure-functions-language-worker-protobuf/src/proto/FunctionRpc.proto",
+    "genTs": "pbjs -t static-module azure-functions-language-worker-protobuf/src/proto/FunctionRpc.proto -o azure-functions-language-worker-protobuf/src/rpc_static.js && pbts -o azure-functions-language-worker-protobuf/src/rpc.d.ts azure-functions-language-worker-protobuf/src/rpc_static.js",
     "test": "mocha -r ts-node/register ./test/**/*.ts"
   },
   "files": [
